requirements.txt
apache-airflow==2.7.0
pyspark==3.5.0
psycopg2==2.9.7
pandas==2.1.0
SQLAlchemy==2.0.20
apache-airflow-providers-postgres==3.5.1